+++ ts/src/workspace.ts
@@ -1,6 +1,8 @@
 import camelCase from "camelcase";
+import * as toml from "toml";
 import { PublicKey } from "@solana/web3.js";
 import { Program } from "./program";
+import { Idl } from "./idl";
 
 let _populatedWorkspace = false;
 
@@ -38,18 +40,36 @@ const workspace = new Proxy({} as any, {
         throw new Error("Could not find workspace root.");
       }
 
+      const idlMap = new Map<string, Idl>();
+
       find
         .fileSync(/target\/idl\/.*\.json/, projectRoot)
         .reduce((programs: any, path: string) => {
           const idlStr = fs.readFileSync(path);
           const idl = JSON.parse(idlStr);
+          idlMap.set(idl.name, idl);
           const name = camelCase(idl.name, { pascalCase: true });
-          programs[name] = new Program(
-            idl,
-            new PublicKey(idl.metadata.address)
-          );
+          if (idl.metadata && idl.metadata.address) {
+            programs[name] = new Program(
+              idl,
+              new PublicKey(idl.metadata.address)
+            );
+          }
           return programs;
         }, workspaceCache);
+
+      // Override the workspace programs if the user put them in the config.
+      const anchorToml = toml.parse(
+        fs.readFileSync(path.join(projectRoot, "Anchor.toml"), "utf-8")
+      );
+      const clusterId = anchorToml.provider.cluster;
+      if (anchorToml.clusters && anchorToml.clusters[clusterId]) {
+        attachWorkspaceOverride(
+          workspaceCache,
+          anchorToml.clusters[clusterId],
+          idlMap
+        );
+      }
 
       _populatedWorkspace = true;
     }
@@ -57,5 +77,20 @@ const workspace = new Proxy({} as any, {
     return workspaceCache[programName];
   },
 });
+
+function attachWorkspaceOverride(
+  workspaceCache: { [key: string]: Program },
+  overrideConfig: { [key: string]: string },
+  idlMap: Map<string, Idl>
+) {
+  Object.keys(overrideConfig).forEach((programName) => {
+    const wsProgramName = camelCase(programName, { pascalCase: true });
+    const overrideAddress = new PublicKey(overrideConfig[programName]);
+    workspaceCache[wsProgramName] = new Program(
+      idlMap.get(programName),
+      overrideAddress
+    );
+  });
+}
 
 export default workspace;